requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
pillow==10.1.0
matplotlib==3.8.2
pydantic==2.5.0
jinja2==3.1.2
python-dotenv==1.0.0
aiofiles==23.2.1
